mtd-nand: DaVinci: Add 4-bit ECC support for 4K page NAND chips (temporary)
This is a temporary patch to add support for 4K page NAND chips. It sets up
the nand_ecclayout structure for 4K page NAND chips and updates eccpos[]
which may have an impact for some of the mtd-utils IOCTLs.

We have used flash_eraseall, nanddump and nandwrite without any issues.

Signed-off-by: Sneha Narnakaje <nsnehaprabha@ti.com>
Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>