V4L : vpif display - updates to support vpif capture on DM6467
The structure name for vpif display driver changed since it was not unique.
So this update is done to reflect the same. Also removed the code related to
register address space iomap. Uses v4l2_i2c_new_subdev_board() instead of
v4l2_i2c_new_probed_subdev() so that platform data can be added for subdevice
configuration for polarities.

This has incorporated comments against version v0 of the patch series.

NOTE: This patch is dependent on the patch from Chaithrika for vpif display.

Reviewed-by: Hans Verkuil <hverkuil@xs4all.nl>

Signed-off-by: Muralidharan Karicheri <m-karicheri2@ti.com>