DaVinci - adding separate platform device for ccdc
This patch is for removing the CCDC resource from vpfe capture
driver and making ccdc a platform driver with its own resources.
This is needed since CCDC has multiple io address space for DM365
and the current implementation is not scalable for extension to
DM365. Also removed Pin MUX configuration to ccdc driver.

NOTE:This patch is meant only for internal release. This patch will be
submitted later for community.

Signed-off-by: Muralidharan Karicheri <m-karicheri2@ti.com>