Adding Digital TV standards for internal use
Resending it... Please use this...

This patch adds Digital TV standards definition for video drivers.
This is only a temporary file used for internal driver usage and
will be removed once support for these standards are added
to V4L2 sub system. Defining it in a separate header file will
ease rebase/merge of upstream kernel to Arago.

NOTE: for internal purpose only !!!

Signed-off-by: Muralidharan Karicheri <m-karicheri2@ti.com>