V4L: Adding ths7353 video amplifier support
This patch add support for THS7353 video amplifier which is used
in the video input path of component inputs. This is similar
to THS7303 driver and is implemented as a v4l2 sub device.
We might want to share the code with ths7303, but didn't find
a way to identify the chip.

NOTE1: Can't submit this to v4l community since we use videohd.h
for digital standard definition.

NOTE2: depends on videohd.h patch that I have sent earlier.

Signed-off-by: Muralidharan Karicheri <m-karicheri2@ti.com>

This patch also fixes the default filter value to be NTSC, instead of using
the NTSC standard value.

Signed-off-by: Sneha Narnakaje <nsnehaprabha@ti.com>