DaVinci: THS7303: Exporting function for HD display
This patch should be considered a hack to get HD display working
on DM365. This will be reverted when the V4L2 display driver is
converted to the subdev model.

This patch also checks for the ths7303 client handle before setting the
filter mode. When the Encoder Manager output is set to COMPONENT in bootargs,
Encoder Manager tries to configure the video display to COMPONENT and default
mode to 720P-60. Since THS7303 driver is not initialized the i2c client handle
is NULL. So a check for NULL was required before setting the filter mode.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>