da830/omap-l137: Reconfigure MSTPRI2 register to increase LCDC priority
da830/omap-l137: Reconfigure MSTPRI2 register to increase LCDC priority

LCDC traffic is typically real-time sensitive, therefore, the
default priority of 5, which is lower as compared to the default
priority of several masters, is not recommended. You should
reconfigure the LCDC priority to the highest or equal to other
high-priority masters in an application to ensure that the
throughput/latency requirements for the LCDC are met.

Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>