davinci: da830/omap-l137: do not force write through cache mode
Silicon revision 2.0 of DA850/OMAP-L137 fixes the advisory forcing
the use of ARM cache in write-through mode.

To fully support the capabilities of the new silicon, stop enforcing
write-through cache mode whenever DA830 is selected. So, users of
new silicon can run it in cache write-back mode.

However, keep cache-writethrough mode enabled in the defconfig so
users of old silicon dont get a nasty surprise when they try to boot.