da850/omapl138: cpufreq: New platform data for PLL async rate
Added a new member, asyncrate, to struct davinci_cpufreq_config,
to facilitate the specification for maximum async frequency to be
used for the platform. This would be used, to (re-)configure the async
related PLL dividers, when frequency change occurs as a result of
DVFS.

Without this change, the async freq would be configured based on
the "previous" divider settings, which may not actually be the
maximum frequency that could be set for async.

Signed-off-by: Madhvapathi Sriram <sriram.m@ti.com>